arao branch update

+++ app/controllers/account/AccountController.php
@@ -50,9 +50,9 @@ class AccountController extends BaseController {
      * Displays the form for cloud account creation
      *
      */
-    public function getCreate($account = false) {
-        $mode = $account !== false ? 'edit' : 'create';
-        $account = $account !== false ? CloudAccount::where('user_id', Auth::id())->findOrFail($account->id) : null;
+    public function getCreate($id = false) {
+        $mode = $id !== false ? 'edit' : 'create';
+		$account =  $id !== false ? CloudAccount::where('user_id', Auth::id())->findOrFail($id) : null;
         $providers = Config::get('account_schema');
         return View::make('site/account/create_edit', compact('mode', 'account', 'providers'));
     }
@@ -94,6 +94,16 @@ class AccountController extends BaseController {
      *
      */
     public function postDelete($account) {
+    		
+    	if(empty($account->id)) $account = CloudAccount::where('user_id', Auth::id())->find($account);
+		
+		$deployment = Deployment::where('user_id', Auth::id())->where('cloudAccountId', $account->id)->get();
+		if(!$deployment->isEmpty())
+		{
+			  return Redirect::to('account')->with('error', 'Deployment is linked with this account and hence cannot be deleted');
+		}
+		
+		
         CloudAccount::where('id', $account->id)->where('user_id', Auth::id())->delete();
         
         $id = $account->id;
